+++ app/views/school_groups/impact/configurations/_form.html.erb
@@ -1,5 +1,5 @@
 <div class='p-2'>
-  <%= simple_form_for(config, url: admin_impact_report_path(config.school_group), method: :patch) do |f| %>
+  <%= simple_form_for(config, url: school_group_impact_configuration_path(config.school_group), method: :patch) do |f| %>
     <%= render Layout::GridComponent.new(cols: 1, cell_classes: 'mb-2 bg-light rounded-2 px-3') do |grid| %>
       <%= grid.with_cell do |row| %>
         <div class="d-flex justify-content-between align-items-center">
